requirements.txt
# Core dependencies
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Firebase
firebase-admin==6.3.0
firebase-functions==0.4.0

# LangChain và AI
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13

# Google AI
google-generativeai==0.3.2

# Vector Database
chromadb==0.4.22

# Utilities
Werkzeug==3.0.1
pandas>=2.2.0
rank-bm25==0.2.2
pytz==2024.1
